CtrlCore: ImageWin32 simplified, RectTracker now works correctly with negative minsize
git-svn-id: svn://ultimatepp.org/upp/trunk@14331 f0d560ea-af0d-0410-9eb7-867de7ffcac7

+++ uppsrc/Core/Util.cpp
@@ -209,7 +209,7 @@ int msecs(int prev)
 
 void TimeStop::Reset()
 {
-	starttime = usecs();
+	starttime = (double)usecs();
 }
 
 TimeStop::TimeStop()
